report/requirements.txt
python-dotenv
pylint
pytest
psycopg2
pandas
ipykernel
